create contractutil
