+++ core/services/ocr2/plugins/ccip/plugins_common.go
@@ -1,28 +1,21 @@
 package ccip
 
 import (
-	"context"
 	"encoding/hex"
 	"math/big"
 	"sort"
 
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/pkg/errors"
 	"golang.org/x/exp/constraints"
 
-	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/client"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/commit_store"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_offramp"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_onramp"
-	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_onramp_1_0_0"
-	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_onramp_1_1_0"
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
-	ccipconfig "github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/config"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/hashlib"
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/observability"
 	"github.com/smartcontractkit/chainlink/v2/core/services/pg"
 	"github.com/smartcontractkit/chainlink/v2/core/utils"
 )
@@ -35,85 +28,6 @@ const (
 )
 
 var ErrCommitStoreIsDown = errors.New("commitStore is down")
-
-func LoadOnRamp(onRampAddress common.Address, pluginName string, client client.Client) (evm_2_evm_onramp.EVM2EVMOnRampInterface, error) {
-	err := ccipconfig.VerifyTypeAndVersion(onRampAddress, client, ccipconfig.EVM2EVMOnRamp)
-	if err != nil {
-		return nil, errors.Wrap(err, "Invalid onRamp contract")
-	}
-	return observability.NewObservedEvm2EvmOnRamp(onRampAddress, pluginName, client)
-}
-
-func LoadOnRampDynamicConfig(onRamp evm_2_evm_onramp.EVM2EVMOnRampInterface, client client.Client) (evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig, error) {
-	versionString, err := onRamp.TypeAndVersion(&bind.CallOpts{})
-	if err != nil {
-		return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{}, err
-	}
-
-	_, version, err := ccipconfig.ParseTypeAndVersion(versionString)
-	if err != nil {
-		return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{}, err
-	}
-
-	opts := &bind.CallOpts{}
-
-	switch version {
-	case "1.0.0":
-		legacyOnramp, err := evm_2_evm_onramp_1_0_0.NewEVM2EVMOnRamp(onRamp.Address(), client)
-		if err != nil {
-			return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{}, err
-		}
-		legacyDynamicConfig, err := legacyOnramp.GetDynamicConfig(opts)
-		if err != nil {
-			return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{}, err
-		}
-		return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{
-			Router:          legacyDynamicConfig.Router,
-			MaxTokensLength: legacyDynamicConfig.MaxTokensLength,
-			PriceRegistry:   legacyDynamicConfig.PriceRegistry,
-			MaxDataSize:     legacyDynamicConfig.MaxDataSize,
-			MaxGasLimit:     uint32(legacyDynamicConfig.MaxGasLimit),
-		}, nil
-	case "1.1.0":
-		legacyOnramp, err := evm_2_evm_onramp_1_1_0.NewEVM2EVMOnRamp(onRamp.Address(), client)
-		if err != nil {
-			return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{}, err
-		}
-		legacyDynamicConfig, err := legacyOnramp.GetDynamicConfig(opts)
-		if err != nil {
-			return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{}, err
-		}
-		return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{
-			Router:                legacyDynamicConfig.Router,
-			MaxTokensLength:       legacyDynamicConfig.MaxTokensLength,
-			DestGasOverhead:       legacyDynamicConfig.DestGasOverhead,
-			DestGasPerPayloadByte: legacyDynamicConfig.DestGasPerPayloadByte,
-			PriceRegistry:         legacyDynamicConfig.PriceRegistry,
-			MaxDataSize:           legacyDynamicConfig.MaxDataSize,
-			MaxGasLimit:           uint32(legacyDynamicConfig.MaxGasLimit),
-		}, nil
-	case "1.2.0":
-		return onRamp.GetDynamicConfig(opts)
-	default:
-		return evm_2_evm_onramp.EVM2EVMOnRampDynamicConfig{}, errors.Errorf("Invalid onramp version: %s", version)
-	}
-}
-
-func LoadOffRamp(offRampAddress common.Address, pluginName string, client client.Client) (evm_2_evm_offramp.EVM2EVMOffRampInterface, error) {
-	err := ccipconfig.VerifyTypeAndVersion(offRampAddress, client, ccipconfig.EVM2EVMOffRamp)
-	if err != nil {
-		return nil, errors.Wrap(err, "Invalid offRamp contract")
-	}
-	return observability.NewObservedEvm2EvmOffRamp(offRampAddress, pluginName, client)
-}
-
-func LoadCommitStore(commitStoreAddress common.Address, pluginName string, client client.Client) (commit_store.CommitStoreInterface, error) {
-	err := ccipconfig.VerifyTypeAndVersion(commitStoreAddress, client, ccipconfig.CommitStore)
-	if err != nil {
-		return nil, errors.Wrap(err, "Invalid commitStore contract")
-	}
-	return observability.NewObservedCommitStore(commitStoreAddress, pluginName, client)
-}
 
 func contiguousReqs(lggr logger.Logger, min, max uint64, seqNrs []uint64) bool {
 	if int(max-min+1) != len(seqNrs) {
@@ -160,17 +74,6 @@ func leavesFromIntervals(
 	}
 
 	return leaves, nil
-}
-
-// Checks whether the commit store is down by doing an onchain check for Paused and ARM status
-func isCommitStoreDownNow(ctx context.Context, lggr logger.Logger, commitStore commit_store.CommitStoreInterface) bool {
-	unPausedAndHealthy, err := commitStore.IsUnpausedAndARMHealthy(&bind.CallOpts{Context: ctx})
-	if err != nil {
-		// If we cannot read the state, assume the worst
-		lggr.Errorw("Unable to read CommitStore IsUnpausedAndARMHealthy", "err", err)
-		return true
-	}
-	return !unPausedAndHealthy
 }
 
 func filterContainsZeroAddress(addrs []common.Address) bool {
